requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
pyodbc==5.0.1
python-dotenv==1.0.0
werkzeug==3.0.1
gunicorn==21.2.0
